This is what i have done during the first part of class
I have added selecting random questions and code to make sure the same question is not selected twice and saves the same answer question and options. I also added lists to store the chosen questions. I also added a loop at the bottom to display the questions and their answers so I know they all like up the the right amount of questions are selected